+++ scheduler/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+* Add `RestartableDescriptor` that can be manually restarted externally to the scheduler lifecycle.
+
 ## [0.2.1] - 2021-02-24
 
 ### Build